chore: Move compare_uuids out of mojang.py
